% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{Analysis of Variance}
\usage{
anova(data, dependent.var, group.var)
}
\arguments{
\item{data}{The data frame to evaluate.}

\item{dependent.var}{Dependent variable of the data frame.}

\item{group.var}{Grouping variable of the data frame. Can be greater 2.}
}
\description{
This function calculates an ANOVA and relevant supplementary statistics.
}
\examples{
carsp <- data.frame(speed = cars$speed, group = as.factor(sample(1:3,nrow(cars),replace = TRUE)))
results.carsp <- proxi::independentt(carsp, speed, group)
}
\keyword{ANOVA}
\keyword{analysis}
\keyword{variance}

