% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentt.R
\name{dependentt}
\alias{dependentt}
\title{Dependent t-test}
\usage{
dependentt(data, dependent.var1, dependent.var2)
}
\arguments{
\item{data}{The data frame to evaluate.}

\item{dependent.var1}{First measurement variable of the data frame.}

\item{dependent.var2}{Second measurement variable of the data frame.}
}
\description{
This function calculates a dependent t-test and relevant supplementary statistics.
}
\examples{
carsp <- data.frame(speed1 = cars$speed[1:(nrow(cars)/2)], speed2 = cars$speed[(nrow(cars)/2+1):nrow(cars)])
results.carsp <- proxi::dependentt(carsp, speed1, speed2)
}
\keyword{dependent}
\keyword{t-test}

