% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentt.R
\name{independentt}
\alias{independentt}
\title{Independent t-test}
\usage{
independentt(data, dependent.var, group.var)
}
\arguments{
\item{data}{The data frame to evaluate.}

\item{dependent.var}{Dependent variable of the data frame.}

\item{group.var}{Grouping variable of the data frame.}
}
\description{
This function calculates an independent t-test and relevant supplementary statistics.
}
\examples{
carsp <- data.frame(speed = cars$speed, group = as.factor(sample(1:2,nrow(cars),replace = TRUE)))
results.carsp <- proxi::independentt(carsp, speed, group)
}
\keyword{independent}
\keyword{t-test}

